Activities done, added & renamed
