+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools">
+    xmlns:tools="http://schemas.android.com/tools" >
 
     <application
         android:allowBackup="true"
@@ -11,7 +11,10 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/Theme.CardGambit"
-        tools:targetApi="31">
+        tools:targetApi="31" >
+        <activity
+            android:name=".GameBoardAI"
+            android:exported="false" />
         <activity
             android:name=".Result"
             android:exported="false" />
@@ -19,14 +22,14 @@
             android:name=".StartGame"
             android:exported="false" />
         <activity
-            android:name=".GameBoard"
-            android:exported="false"/>
+            android:name=".GameBoardFriend"
+            android:exported="false" />
         <activity
             android:name=".Instructions"
             android:exported="false" />
         <activity
             android:name=".FrontPage"
-            android:exported="true">
+            android:exported="true" >
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
