+++ app/src/main/res/layout/activity_game_board_ai.xml
@@ -5,12 +5,6 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".GameBoard">
+    tools:context=".GameBoardAI">
 
-    <com.google.android.material.button.MaterialButton
-        android:text="Instructions"
-        android:id="@+id/btn_instructions"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:backgroundTint="@color/blue"/>
 </androidx.constraintlayout.widget.ConstraintLayout>
