+++ app/src/main/res/layout/activity_instructions.xml
@@ -7,10 +7,34 @@
     android:layout_height="match_parent"
     tools:context=".Instructions">
 
+<LinearLayout
+    android:layout_width="wrap_content"
+    android:layout_height="wrap_content"
+    android:orientation="vertical">
+
+    <ImageView
+        android:id="@+id/bridgeInstructions"
+        android:layout_width="wrap_content"
+        android:layout_height="300dp"
+        app:srcCompat="@drawable/bridgecards"
+        tools:layout_editor_absoluteX="0dp"
+        tools:layout_editor_absoluteY="-0dp" />
+    <ImageView
+        android:id="@+id/backesideInstructions"
+        android:layout_width="wrap_content"
+        android:layout_height="300dp"
+        app:srcCompat="@drawable/backsidecard"
+        tools:layout_editor_absoluteX="0dp"
+        tools:layout_editor_absoluteY="40dp" />
+
     <com.google.android.material.button.MaterialButton
-        android:text="result"
-        android:id="@+id/btn_result"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:backgroundTint="@color/cyan"/>
+    android:id="@+id/btn_startgame"
+    android:layout_width="match_parent"
+    android:layout_height="wrap_content"
+    android:layout_marginBottom="16dp"
+    android:backgroundTint="@color/red"
+    android:text="Start Game"
+    app:layout_constraintBottom_toBottomOf="parent"
+    tools:layout_editor_absoluteX="0dp" />
+</LinearLayout>
 </androidx.constraintlayout.widget.ConstraintLayout>
